+++ src/cli.rs
@@ -176,9 +176,6 @@ enum Commands {
         /// Skip lint check
         #[arg(long)]
         no_lint: bool,
-        /// Disable JSON output to stdout
-        #[arg(long)]
-        no_json: bool,
         /// Disable ::error::/::warning:: annotations
         #[arg(long)]
         no_annotations: bool,
@@ -279,7 +276,6 @@ async fn main() -> ExitCode {
             no_owners_changed,
             no_owners_all,
             no_lint,
-            no_json,
             no_annotations,
             no_summary,
             no_outputs,
@@ -303,7 +299,6 @@ async fn main() -> ExitCode {
                 check_owners_changed: !no_owners_changed,
                 check_owners_all: !no_owners_all,
                 check_lint: !no_lint,
-                output_json: !no_json,
                 output_annotations: !no_annotations,
                 output_summary: !no_summary,
                 output_vars: !no_outputs,